Gerant + Year

+++ src/app/client/clients/clients.component.ts
@@ -3,6 +3,7 @@ import { Client } from '../../model/client.model';
 import { ClientService } from '../../service/client.service';
 import { AuthentificationService } from 'src/app/service/authentification.service';
 import {Year} from "../../model/year.model";
+import {YearService} from "../../service/year.service";
 @Component({
   selector: 'app-clients',
   templateUrl: './clients.component.html',
@@ -17,28 +18,30 @@ export class ClientsComponent implements OnInit {
   newIdYear! : number;
   newYear! : Year;
 
-  constructor(private clientService : ClientService, public authService : AuthentificationService) { }
+  constructor(private clientService : ClientService, public authService : AuthentificationService,
+              public yearService : YearService) { }
 
   ngOnInit(): void {
     this.chargerClients();
-    /*this.clientService.listYears(). subscribe(year => {
-      this.years = year;
-    });*/
+    this.chargerYears();
   }
 
   chargerClients(){
-    //this.newClient.year = this.years.find(y => y.idYear == this.newIdYear)!;
     this.clientService.listeClient().subscribe(client => {
       this.clients = client;
     });
   }
-
+  chargerYears(){
+    this.yearService.listeYears(). subscribe(year => {
+      this.years = year;
+    });
+  }
   supprimerClient(client : Client) {
-    //let conf = confirm("Etes-vous sûr ?");
-   // if (conf) {
+    let conf = confirm("Etes-vous sûr ?");
+    if (conf) {
       this.clientService.supprimerClient(client.idClient).subscribe(() => {
         this.chargerClients();
       });
-    //}
+    }
   }
 }
